+++ test/CosmosDB.utils.Tests.ps1
@@ -12,6 +12,7 @@ InModuleScope CosmosDB {
     # Variables for use in tests
     $script:testAccount = 'testAccount'
     $script:testDatabase = 'testDatabase'
+    $script:testCollection = 'testCollection'
     $script:testKey = 'GFJqJeri2Rq910E0G7PsWoZkzowzbj23Sm9DUWFC0l0P8o16mYyuaZKN00Nbtj9F1QQnumzZKSGZwknXGERrlA=='
     $script:testKeySecureString = ConvertTo-SecureString -String $script:testKey -AsPlainText -Force
     $script:testEmulatorKey = 'C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw=='
@@ -25,6 +26,21 @@ InModuleScope CosmosDB {
         KeyType  = 'master'
         BaseUri  = ('https://{0}.documents.azure.com/' -f $script:testAccount)
     }
+    $script:testToken = 'type-resource&ver=1.0&sig=5mDuQBYA0kb70WDJoTUzSBMTG3owkC0/cEN4fqa18/s='
+    $script:testTokenSecureString = ConvertTo-SecureString -String $script:testToken -AsPlainText -Force
+    $script:testTokenResource = ('dbs/{0}/colls/{1}' -f $script:testDatabase, $script:testCollection)
+    $script:testContextToken = [CosmosDB.ContextToken] @{
+        Resource  = $script:testTokenResource
+        TimeStamp = $script:testDate
+        Token     = $script:testTokenSecureString
+    }
+    $script:testResourceContext = [CosmosDb.Context] @{
+        Account  = $script:testAccount
+        Database = $script:testDatabase
+        BaseUri  = ('https://{0}.documents.azure.com/' -f $script:testAccount)
+        Token    = $script:testContextToken
+    }
+
     $script:testJson = @'
 {
     "_rid": "2MFbAA==",
@@ -71,7 +87,7 @@ InModuleScope CosmosDB {
             { Get-Command -Name New-CosmosDbContext -ErrorAction Stop } | Should -Not -Throw
         }
 
-        Context 'When called with Context parameters' {
+        Context 'When called with Account parameters' {
             $script:result = $null
 
             It 'Should not throw exception' {
@@ -94,7 +110,7 @@ InModuleScope CosmosDB {
             }
         }
 
-        Context 'When called with Azure parameters and not connected to Azure' {
+        Context 'When called with AzureAccount parameters and not connected to Azure' {
             $script:result = $null
 
             Mock -CommandName Get-AzureRmContext -MockWith { throw }
@@ -132,7 +148,7 @@ InModuleScope CosmosDB {
             }
         }
 
-        Context 'When called with Azure parameters and connected to Azure' {
+        Context 'When called with AzureAccount parameters and connected to Azure' {
             $script:result = $null
 
             Mock `
@@ -144,7 +160,7 @@ InModuleScope CosmosDB {
                     secondaryReadonlyMasterKey = 'secondaryReadonlyMasterKey'
                 } }
 
-            Mock -CommandName Get-AzureRmContext -MockWith { $true  }
+            Mock -CommandName Get-AzureRmContext -MockWith { $true }
             Mock -CommandName Add-AzureRmAccount
 
             It 'Should not throw exception' {
@@ -178,6 +194,7 @@ InModuleScope CosmosDB {
                 $newCosmosDbContextParameters = @{
                     Database = $script:testDatabase
                     Emulator = $true
+                    Verbose  = $true
                 }
 
                 { $script:result = New-CosmosDbContext @newCosmosDbContextParameters } | Should -Not -Throw
@@ -190,6 +207,59 @@ InModuleScope CosmosDB {
                 $tempCredential.Password | Should -Be $script:testEmulatorKey
                 $script:result.KeyType | Should -Be 'master'
                 $script:result.BaseUri | Should -Be ('https://localhost:8081/')
+            }
+        }
+
+        Context 'When called with Token parameters' {
+            $script:result = $null
+
+            It 'Should not throw exception' {
+                $newCosmosDbContextParameters = @{
+                    Account  = $script:testAccount
+                    Database = $script:testDatabase
+                    Token    = $script:testContextToken
+                    Verbose  = $true
+                }
+
+                { $script:result = New-CosmosDbContext @newCosmosDbContextParameters } | Should -Not -Throw
+            }
+
+            It 'Should return expected result' {
+                $script:result.Account | Should -Be $script:testAccount
+                $script:result.Database | Should -Be $script:testDatabase
+                $script:result.BaseUri | Should -Be ('https://{0}.documents.azure.com/' -f $script:testAccount)
+                $script:result.Token[0].Resource | Should -Be $script:testTokenResource
+                $script:result.Token[0].TimeStamp | Should -Be $script:testDate
+                $BSTR = [System.Runtime.InteropServices.Marshal]::SecureStringToBSTR($script:result.Token[0].Token)
+                $decryptedToken = [System.Runtime.InteropServices.Marshal]::PtrToStringAuto($BSTR)
+                $decryptedToken | Should -Be $script:testToken
+            }
+        }
+    }
+
+    Describe 'New-CosmosDbContextToken' -Tag 'Unit' {
+        It 'Should exist' {
+            { Get-Command -Name New-CosmosDbContextToken -ErrorAction Stop } | Should -Not -Throw
+        }
+
+        Context 'When called with paramters' {
+            $script:result = $null
+
+            It 'Should not throw exception' {
+                $newCosmosDbContextTokenParameters = @{
+                    Resource  = $script:testTokenResource
+                    TimeStamp = $script:testDate
+                    Token     = $script:testTokenSecureString
+                    Verbose   = $true
+                }
+
+                { $script:result = New-CosmosDbContextToken @newCosmosDbContextTokenParameters } | Should -Not -Throw
+            }
+
+            It 'Should return expected result' {
+                $script:result.Resource | Should -Be $script:testTokenResource
+                $script:result.TimeStamp | Should -Be $script:testDate
+                $script:result.Token | Should -Be $script:testTokenSecureString
             }
         }
     }
@@ -205,6 +275,7 @@ InModuleScope CosmosDB {
             It 'Should not throw exception' {
                 $GetCosmosDbUriParameters = @{
                     Account = $script:testAccount
+                    Verbose = $true
                 }
 
                 { $script:result = Get-CosmosDbUri @GetCosmosDbUriParameters } | Should -Not -Throw
@@ -223,6 +294,7 @@ InModuleScope CosmosDB {
                 $GetCosmosDbUriParameters = @{
                     Account = $script:testAccount
                     BaseUri = $script:testBaseUri
+                    Verbose = $true
                 }
 
                 { $script:result = Get-CosmosDbUri @GetCosmosDbUriParameters } | Should -Not -Throw
@@ -245,7 +317,8 @@ InModuleScope CosmosDB {
 
             It 'Should not throw exception' {
                 $convertToCosmosDBTokenDateStringParameters = @{
-                    Date = $script:testDate
+                    Date    = $script:testDate
+                    Verbose = $true
                 }
 
                 { $script:result = ConvertTo-CosmosDbTokenDateString @convertToCosmosDBTokenDateStringParameters } | Should -Not -Throw
@@ -273,6 +346,7 @@ InModuleScope CosmosDB {
                     ResourceType = 'users'
                     ResourceId   = 'dbs/testdb'
                     Date         = $script:testUniversalDate
+                    Verbose      = $true
                 }
 
                 { $script:result = New-CosmosDbAuthorizationToken @newCosmosDbAuthorizationTokenParameters } | Should -Not -Throw
@@ -294,6 +368,7 @@ InModuleScope CosmosDB {
                     ResourceType = 'users'
                     ResourceId   = 'dbs/Testdb'
                     Date         = $script:testUniversalDate
+                    Verbose      = $true
                 }
 
                 { $script:result = New-CosmosDbAuthorizationToken @newCosmosDbAuthorizationTokenParameters } | Should -Not -Throw
@@ -322,7 +397,7 @@ InModuleScope CosmosDB {
             $invokeRestMethod_parameterfilter = {
                 $Method -eq 'Get' -and `
                 $ContentType -eq 'application/json' -and `
-                $Uri -eq ('{0}dbs/{1}/{2}' -f $script:testContext.BaseUri,$script:testContext.Database,'users')
+                $Uri -eq ('{0}dbs/{1}/{2}' -f $script:testContext.BaseUri, $script:testContext.Database, 'users')
             }
 
             Mock `
@@ -337,6 +412,7 @@ InModuleScope CosmosDB {
                     Context      = $script:testContext
                     Method       = 'Get'
                     ResourceType = 'users'
+                    Verbose      = $true
                 }
 
                 { $script:result = Invoke-CosmosDbRequest @invokeCosmosDbRequestparameters } | Should -Not -Throw
@@ -358,8 +434,8 @@ InModuleScope CosmosDB {
         Context 'When called with context parameter and Get method and ResourceType is ''dbs''' {
             $invokeRestMethod_parameterfilter = {
                 $Method -eq 'Get' -and `
-                $ContentType -eq 'application/json' -and `
-                $Uri -eq ('{0}{1}' -f $script:testContext.BaseUri,'dbs')
+                    $ContentType -eq 'application/json' -and `
+                    $Uri -eq ('{0}{1}' -f $script:testContext.BaseUri, 'dbs')
             }
 
             Mock `
@@ -374,6 +450,7 @@ InModuleScope CosmosDB {
                     Context      = $script:testContext
                     Method       = 'Get'
                     ResourceType = 'dbs'
+                    Verbose      = $true
                 }
 
                 { $script:result = Invoke-CosmosDbRequest @invokeCosmosDbRequestparameters } | Should -Not -Throw
@@ -395,8 +472,8 @@ InModuleScope CosmosDB {
         Context 'When called with context parameter and Get method and ResourceType is ''offers''' {
             $invokeRestMethod_parameterfilter = {
                 $Method -eq 'Get' -and `
-                $ContentType -eq 'application/json' -and `
-                $Uri -eq ('{0}{1}' -f $script:testContext.BaseUri,'offers')
+                    $ContentType -eq 'application/json' -and `
+                    $Uri -eq ('{0}{1}' -f $script:testContext.BaseUri, 'offers')
             }
 
             Mock `
@@ -411,6 +488,7 @@ InModuleScope CosmosDB {
                     Context      = $script:testContext
                     Method       = 'Get'
                     ResourceType = 'offers'
+                    Verbose      = $true
                 }
 
                 { $script:result = Invoke-CosmosDbRequest @invokeCosmosDbRequestparameters } | Should -Not -Throw
@@ -429,11 +507,88 @@ InModuleScope CosmosDB {
             }
         }
 
+        Context 'When called with context parameter and Get method and ResourceType is ''colls''' {
+            $invokeRestMethod_parameterfilter = {
+                $Method -eq 'Get' -and `
+                $ContentType -eq 'application/json' -and `
+                $Uri -eq ('{0}dbs/{1}/{2}' -f $script:testContext.BaseUri, $script:testContext.Database, 'colls')
+            }
+
+            Mock `
+                -CommandName Invoke-RestMethod `
+                -ParameterFilter $invokeRestMethod_parameterfilter `
+                -MockWith $invokeRestMethod_mockwith
+
+            $script:result = $null
+
+            It 'Should not throw exception' {
+                $invokeCosmosDbRequestparameters = @{
+                    Context      = $script:testContext
+                    Method       = 'Get'
+                    ResourceType = 'colls'
+                    Verbose      = $true
+                }
+
+                { $script:result = Invoke-CosmosDbRequest @invokeCosmosDbRequestparameters } | Should -Not -Throw
+            }
+
+            It 'Should return expected result' {
+                $script:result._count | Should -Be 1
+            }
+
+            It 'Should call expected mocks' {
+                Assert-MockCalled `
+                    -CommandName Invoke-RestMethod `
+                    -ParameterFilter $invokeRestMethod_parameterfilter `
+                    -Exactly -Times 1
+                Assert-MockCalled -CommandName Get-Date -Exactly -Times 1
+            }
+        }
+
+        Context 'When called with resource context parameter and Get method and ResourceType is ''colls''' {
+            $invokeRestMethod_parameterfilter = {
+                $Method -eq 'Get' -and `
+                $ContentType -eq 'application/json' -and `
+                $Uri -eq ('{0}dbs/{1}/colls/{2}' -f $script:testContext.BaseUri, $script:testContext.Database, $script:testCollection)
+            }
+
+            Mock `
+                -CommandName Invoke-RestMethod `
+                -ParameterFilter $invokeRestMethod_parameterfilter `
+                -MockWith $invokeRestMethod_mockwith
+
+            $script:result = $null
+
+            It 'Should not throw exception' {
+                $invokeCosmosDbRequestparameters = @{
+                    Context      = $script:testResourceContext
+                    Method       = 'Get'
+                    ResourceType = 'colls'
+                    ResourcePath = ('colls/{0}' -f $script:testCollection)
+                    Verbose      = $true
+                }
+
+                { $script:result = Invoke-CosmosDbRequest @invokeCosmosDbRequestparameters } | Should -Not -Throw
+            }
+
+            It 'Should return expected result' {
+                $script:result._count | Should -Be 1
+            }
+
+            It 'Should call expected mocks' {
+                Assert-MockCalled `
+                    -CommandName Invoke-RestMethod `
+                    -ParameterFilter $invokeRestMethod_parameterfilter `
+                    -Exactly -Times 1
+                Assert-MockCalled -CommandName Get-Date -Exactly -Times 0
+            }
+        }
+
         Context 'When called with context parameter and Post method' {
             $invokeRestMethod_parameterfilter = {
                 $Method -eq 'Post' -and `
                 $ContentType -eq 'application/query+json' -and `
-                $Uri -eq ('{0}dbs/{1}/{2}' -f $script:testContext.BaseUri,$script:testContext.Database,'users')
+                $Uri -eq ('{0}dbs/{1}/{2}' -f $script:testContext.BaseUri, $script:testContext.Database, 'users')
             }
 
             Mock `
@@ -450,6 +605,7 @@ InModuleScope CosmosDB {
                     ResourceType = 'users'
                     ContentType  = 'application/query+json'
                     Body         = '{ "id": "daniel" }'
+                    Verbose      = $true
                 }
 
                 { $script:result = Invoke-CosmosDbRequest @invokeCosmosDbRequestparameters } | Should -Not -Throw